Revert "Move Data.Stream.Async to Data.Stream.IsStream.Async"
This reverts commit 30c98dc65ff821b5332cb4f847716df5b2e0dc0a.
